+++ maps-sdk/src/main/java/com/mapbox/maps/MapboxMap.kt
@@ -2324,6 +2324,35 @@ class MapboxMap :
   }
 
   /**
+   * A convex polygon that describes the shape of the screen in case it is non-rectangular.
+   * Every coordinate is in 0 to 1 range, with (0, 0) being the map view top-left and (1, 1) the bottom-right.
+   * The points have to be given in clockwise winding order. The polygon will be closed automatically, so for a rectangular shape, pass in 4 points.
+   * Use this, if the visible screen area is obscured enough that using a custom shape improves performance.
+   *
+   * @param shape shape of the screen
+   *
+   */
+  @MapboxExperimental
+  fun setScreenCullingShape(shape: List<Vec2>) {
+    checkNativeMap("setScreenCullingShape")
+    return nativeMap.setScreenCullingShape(shape)
+  }
+
+  /**
+   * A convex polygon that describes the shape of the screen in case it is non-rectangular.
+   * Every screen-coordinate is in 0 to 1 range, with (0, 0) being top-left and (1, 1) bottom-right.
+   * The points have to be given in clockwise winding order. The polygon will be closed automatically, so for a rectangular shape, pass in 4 points.
+   * Use this if the visible screen area is obscured enough that using a custom shape improves performance.
+   *
+   * @return shape of the screen
+   */
+  @MapboxExperimental
+  fun getScreenCullingShape(): List<Vec2> {
+    checkNativeMap("getScreenCullingShape")
+    return nativeMap.getScreenCullingShape()
+  }
+
+  /**
    * Enable real-time collection of map rendering performance statistics, for development purposes. Use after `render()` has
    * been called for the first time.
    *